requirements.txt
streamlit>=1.30.0
langgraph>=0.0.25
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.14
openai>=1.0.0
yfinance>=0.2.18
python-dotenv>=1.0.0
openbb>=3.2.0
reportlab>=4.0.8
psycopg2-binary>=2.9.9
schedule>=1.2.0
pytz>=2024.1
requests
pandas
ipython>=8.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
